src/requirements.txt
pika==1.0.1
python-dotenv==0.10.3
